+++ src/instructions/mod.rs
@@ -9,13 +9,6 @@ pub mod segmentation;
 pub mod tables;
 pub mod tlb;
 
-/// Cause a breakpoint exception by invoking the `int3` instruction.
-pub fn int3() {
-    unsafe {
-        asm!("int3" :::: "volatile");
-    }
-}
-
 /// Halts the CPU until the next interrupt arrives.
 #[inline(always)]
 pub fn hlt() {